Effective Credit Risk Management in Finance using Big Data Technology
The project shows the Poster i would be presenting at the University's research symposium for Engineering and Technology summit, taking place on April 21, 2023
This project utilizes the power of Big data using Hadoop and Apache Spark as the storage and processing engine respectively, subsequently, the project helps banks to be able to handle risk management in the disbursement of loans and facilities by using the model to determine the likelihood of defaults. The project classified customers as defaulters and non-defaulters based on their previous records/ Historical data. I thereafter trained my model to learn from patterns using machine learning algorithms such as Decision Trees, Naive Bayes, and Random Forest.  The model was split into test-train-split with the ratio of 70% of the data for training and 30% of the data for testing. To check the validity of my model, I used some performance index such as Accuracy, Precision, F1 score, Recall, and confusion matrix to check the performance of my model. The result gave a very nice output with the accuracy of 90%. The random forest algorithm gave the best results. However, we cannot determine which model would always give the best output. This is because the output of the model is dependent on the type of data being used,